Adding area-weighted calc to shgc of merged typology in Dragonfly
Calculated the new shgc by area-weighting it by it's glazing area (not the facade area, but the actual window area of each building).

Additionally, I made one stylistic move: I moved the roof_albedo, and roof_veg_fraction up so that the calculation of the new value, and assignment of that value is done with the other area-weighted variables.